FIXED: Intermittent unit test hang due to race condition with scheduler.
This change involved moving the 'fire' capability out of the event and into
the scheduler.  Next, the scheduler doesn't unlock the barrier unless the
runtime parser indicates that the barrier should be unlocked.

These changes also affected sample.c.  Sample.c was updated to get it to
compile, however, the testing framework doesn't appear to work, and I didn't
debug further.  I'm not sure if we're moving forward with that framework, but
if so, we can debug it later.